FreshStart
Could not wrap my head around what to do in Unity scripts to get my display object to store its order to list. Found it easier to have display balls be display object as well. Watching a video series on how to make a Simon game by "gamesplusjames" along with a video on Simon by a creator called "Yasil" is helping get my head around it.